tbh5-kitt/requirements.txt
rpaframework==9.*
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
datetime
fitz
PyMuPDF==1.16.14
ta-bitwarden-cli
numpy
